Init something hard

+++ src/main/java/com/takypok/report/model/entity/Priority.java
@@ -12,6 +12,6 @@ import lombok.extern.slf4j.Slf4j;
 @ToString
 public class Priority extends IdEntity {
   private String name;
-  private Long responseTime;
-  private Long resolutionTime;
+  private Double responseTime;
+  private Double resolutionTime;
 }
