+++ src/main/java/com/takypok/report/model/request/PriorityUpdateRequest.java
@@ -11,6 +11,6 @@ import lombok.*;
 public class PriorityUpdateRequest {
   @NotNull private Long id;
   @NotNull private String name;
-  @NotNull private Long responseTime;
-  @NotNull private Long resolutionTime;
+  @NotNull private Double responseTime;
+  @NotNull private Double resolutionTime;
 }
