+++ src/main/resources/db/changelog/init/init-function.sql
@@ -220,8 +220,8 @@ BEGIN
                                 (s.setting ->> 'workEnd')::TIME                                 AS work_end,
                                 (s.setting ->> 'lunchStart')::TIME                              AS lunch_start,
                                 (s.setting ->> 'lunchEnd')::TIME                                AS lunch_end,
-                                (s.priority ->> 'responseTime')::INTEGER                        AS response_time,
-                                (s.priority ->> 'resolutionTime')::INTEGER                      AS resolution_time,
+                                (s.priority ->> 'responseTime')::NUMERIC                        AS response_time,
+                                (s.priority ->> 'resolutionTime')::NUMERIC                      AS resolution_time,
                                 json_to_tstzrange_array(s.paused_time)                          AS paused_ranges,
                                 ARRAY(SELECT jsonb_array_elements(s.setting -> 'weekend')::int) AS weekend_days
                          FROM ticket t
@@ -237,8 +237,8 @@ BEGIN
                                  )
                              )
                            AND NOW() - t.created_at > LEAST(
-                                                              (s.priority ->> 'responseTime')::int,
-                                                              (s.priority ->> 'resolutionTime')::int
+                                                              (s.priority ->> 'responseTime')::numeric,
+                                                              (s.priority ->> 'resolutionTime')::numeric
                                                       ) * interval '1 hour'),
          office_time_data AS (SELECT td.*,
                                      calculate_office_time(
